Import spectral element I/O code
This code provides a model for 2D LGL discontinuous spectral element
grids and a routine for reading them from a custom file format.  It also
implements interpolation under the assumption that the coordinate map is
restricted to translation and scaling.

Code quality is rather poor, but sufficient for the time being.