Add feature allowing renderings to be saved
Renderings of all timesteps can be saved as PNG images in a
user-specified directory by clicking a button.  A message is also
printed on the console providing instructions for assembling the frames
into an animation.

This change also refactors some of the message types, introducing
SequencedMessage as a message wrapper with a sequence number.